+++ lib/awesome_jsonb_translate.rb
@@ -4,6 +4,6 @@ require 'active_record'
 require 'awesome_jsonb_translate/version'
 require 'awesome_jsonb_translate/active_record'
 
+# Provides PostgreSQL JSONB-based translation functionality for ActiveRecord models
 module AwesomeJsonbTranslate
-  # Just serve as a namespace
 end